macos: revise graceful termination approach
Freeing the surfaces directly was a flawed approach because its a
fundamental structure within the application.

Instead, we now gather all of the surfaces and stop their processes. The
surfaces remain alive which allows us to poll them and delay application
termination (up to a time limit) until they've exited.